revert the truthiness of action.emailHash
the code started out as:
```
emailHash = !!action.emailHash ? action.emailHash :
                                 md5(action.email.trim().toLowerCase());
```

then it was changed to: `if (action.email && !!emailHash) {`

the intent is if the email hash does not exist in localStorage
generate a new one, if it exists in localStorage pull it out of
there and don't generate a new one.

after we ran it we found another problem where the localStorage item
would be set to undefined instead of null.  So we changed it so that
if the action.email was undefined it would set the localStorage to null.
All of the other localStorage items are set to null incase they don't
have a value.