progress bar functional when provisioning
This was actually always a problem, it was just found when we
went to a progress bar, now it works as expected.  Tests had to be
changed for an additional call to dispatch.  In addition the code we
were editing was right next to code that controlled creating multiple
linodes at the same time.  The multiple linodes at the same time code
was removed from the view so this commit also removes that section of
code as well.  This can be added back in at a later time by referencing
a commit earlier than this one; if a decision is made to add that
functionality back in.