[ Proposal ] Don't Die on Unrecognized Events
This relates to https://github.com/linode/manager/issues/2621

The api adds event actions with most new features, and unless they are
met immediately with a change here, they cause a "This page is broken"
message.  This is a proposal to display a generic event notification for
unrecognized events.  Ideally this would never happen, but realistically
this could provide a stop-gap solution while proper handling for new
events can be implemented, and should reduce downtime.