Add getDerivedState method to LinodeSettingsLabelPanel
Due to (I think) disparity between initialValue and alteredValue,
changes made to the Linode label from the EditableText at the top
of the view were not reflected in the LabelPanel (the change
was reflected in the other direction, however). I added
getDerivedStateFromProps to make sure that this update would be made.