M3-1008 Fix Image search link (#3703)
* M3-1008 Fix Image search link

If you click on a link for an image search result in the search bar,
the link takes you to an image detail page (which does not exist
and is not currently planned).

This PR causes all image search results to link to the `/images`
landing page, where all images can be viewed in a list.

* Show 404 for image path

Added `exact` to the `<Route>` for images. This way, attempting
to visit /images/private/123455 will show a Not Found screen
rather than a blank page.