OBJ-153 change: Display bucket size using base 2 instead of base 10 (#4914)
* Create unitConversion utility file

Move the existing `convertMegabytesTo` function to a new utility file. Create a new function called `readableBytes`, which takes in a number in bytes and returns a human-readable version.

This change was originally written in the closed PR, #4857. It was cleaner to re-write it instead of cherry picking a partial commit.

* Add tests for readableBytes

* Use base 2 when displaying bucket size