OBJ-138 Change: Use account.capabilities to determine if OBJ is enabled. (#4916)
* Use /profile to determine isObjectStorageEnabled

Previously, OBJ was enabled per environment, set by a variable in .env.

Now, we ALSO check the response on /profile, which may return `beta_programs: ["Object Storage EAP"]`. If a user is enrolled in the Object Storage EAP beta program, we enable OBJ, even if it's globally disabled for the environment.

This allows us to enable OBJ per-customer in a production environment.

* Use fuzzier string matching (review feedback)

* Update CDU logic in PrimaryNav (review feedback)

Check to see if `profile` has been updated, because profile.beta_programs could have changed, which means that we might need to show different menu items.

Also check if the account is managed for the same reason.

* Move OBJ requests to OBJ Landing instead of App

App.tsx fetches `profile`, which is now used to determine if OBJ is enabled. We used to request buckets and clusters on initial App load, but now we need to wait for the profile response.

I could have:
1) Implemented getProfile().then(() => getEverythingElse()) sort of logic.
2) Added CDU logic to App.tsx to check if profile has been loaded, THEN maybe do OBJ requests.
3) Move OBJ requests to ObjectStorageLanding.

I opted for #3, because the first two options resulted in a performance hit for all users.

We could move these requests back to App.tsx once OBJ is generally available.

* Dispatch requestAccount() on App load

* Add Linode.AccountCapability type

This is currently only returned from the beta API, so it is not yet included on the Account interface.

* Use account.capabilities vs. profile.beta_programs

* Add missing PrimaryNav test

* Update accountCapabilities tests

* Fix comment

* Add loading and errors states

We show a loading state if we don’t have a response from /account yet, otherwise there would be a brief 404 while /account loads if you go to /object-storage/buckets directly.

We also handle the request to /account resulting in an error.

* Use React.FC instead of StatelessComponent

* Add LandingLoading component

This component accepts props that control delay of rendering the component. It could potentially be used for everywhere we display <CircleProgress /> on a Landing page.

* Use <LandingLoading /> and simplify Error state